% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_data_country_stats.R
\name{lm_data_country_stats}
\alias{lm_data_country_stats}
\title{Get country stats data for use in report}
\usage{
lm_data_country_stats()
}
\value{
\code{dataframe}.
}
\description{
\code{lm_data_country_stats} returns formatted country stats data for use in monthly reports.
}
\examples{
\dontrun{
lm_data_country_stats()
}

}
