% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_data_key_stats.R
\name{lm_data_key_stats}
\alias{lm_data_key_stats}
\title{Get key stats data for use in report}
\usage{
lm_data_key_stats()
}
\value{
\code{dataframe}.
}
\description{
\code{lm_data_key_stats} returns formatted key stats data for use in monthly reports.
}
\examples{
\dontrun{
lm_data_key_stats()
}

}
