% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_data_nomis_example.R
\name{lm_data_nomis_example}
\alias{lm_data_nomis_example}
\title{Example function calling from NOMIS API, to be deleted}
\usage{
lm_data_nomis_example()
}
\value{
\code{dataframe}.
}
\description{
\code{lm_data_nomis_example} is an example of a function that pulls data from the NOMIS API.
}
\examples{
\dontrun{
lm_data_NOMIS_example()
}

}
