% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_data_ons_example.R
\name{lm_data_ons_example}
\alias{lm_data_ons_example}
\title{Example function calling from ONS API, to be deleted}
\usage{
lm_data_ons_example()
}
\value{
\code{dataframe}.
}
\description{
\code{lm_data_ons_example} is an example of a function that pulls data from the ONS API.
}
\examples{
\dontrun{
lm_data_ons_example()
}

}
