% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_data_region_stats.R
\name{lm_data_region_stats}
\alias{lm_data_region_stats}
\title{Get region stats data for use in report}
\usage{
lm_data_region_stats()
}
\value{
\code{dataframe}.
}
\description{
\code{lm_data_region_stats} returns formatted region stats data for use in monthly reports.
}
\examples{
\dontrun{
lm_data_region_stats()
}

}
