% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_data_time_stats.R
\name{lm_data_time_stats}
\alias{lm_data_time_stats}
\title{Get time stats data for use in report}
\usage{
lm_data_time_stats()
}
\value{
\code{dataframe}.
}
\description{
\code{lm_data_time_stats} returns formatted time stats data for use in monthly reports.
}
\examples{
\dontrun{
lm_data_time_stats()
}

}
