% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_format_get_request.R
\name{lm_format_get_request}
\alias{lm_format_get_request}
\title{Make GET request to API and format response as R list}
\usage{
lm_format_get_request(uri)
}
\arguments{
\item{uri}{\code{string}.
The URI to make the GET request to.}
}
\value{
\code{list}.
}
\description{
\code{lm_format_get_request} sends a GET request to a provided URI and formats the JSON response as an R list.
}
\examples{
\dontrun{
lm_format_get_request(uri = 'https://api.beta.ons.gov.uk/v1/datasets')
}

}
