% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_nomis_dataset_dimension_options.R
\name{lm_nomis_dataset_dimension_options}
\alias{lm_nomis_dataset_dimension_options}
\title{Return a list of options of a dimension of a dataset in the NOMIS API}
\usage{
lm_nomis_dataset_dimension_options(base_uri, dataset_id, dataset_dimension)
}
\arguments{
\item{base_uri}{\code{string}.
The base URI to connect to.}

\item{dataset_id}{\code{string}.
An ID of an existing dataset in the NOMIS API. See \link{lm_nomis_datasets}.}

\item{dataset_dimension}{\code{string}.
The name of an existing dimension in the specified dataset. See \link{lm_nomis_dataset_dimensions}.}
}
\value{
\code{character} list of available options.
}
\description{
\code{lm_ons_dataset_dimension_options} returns a list of options of a dimension of a dataset available on the
   \href{https://www.nomisweb.co.uk/api/v01/help}{NOMIS API}.
}
\examples{
\dontrun{
lm_nomis_dataset_dimension_options(base_uri = "http://www.nomisweb.co.uk/", dataset_id = "NM_1_1", dataset_dimension = "FREQ")
}

}
