% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_nomis_dataset_dimensions.R
\name{lm_nomis_dataset_dimensions}
\alias{lm_nomis_dataset_dimensions}
\title{Return list of dimensions of a dataset in the NOMIS API}
\usage{
lm_nomis_dataset_dimensions(dataset_id, datasets_list)
}
\arguments{
\item{dataset_id}{\code{string}.
An ID matching a dataset included in \code{dataset_list}.}

\item{datasets_list}{\code{list}.
A table of datsets created using \link{lm_nomis_datasets}.}
}
\value{
\code{character} a list of available dimensions in the dataset.
}
\description{
\code{lm_ons_dataset_dimensions} returns a list of dimensions dataset available on the
   \href{https://www.nomisweb.co.uk/api/v01/help}{NOMIS API}.
}
\examples{
\dontrun{
datasets <- lm_nomis_datasets(base_uri = "http://www.nomisweb.co.uk/", variables = c("id", "uri", "description.value", "name.value", "components.dimension"))
lm_nomis_dataset_dimensions(dataset_id = "NM_1_1", datasets_list = datasets)
}

}
