% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_nomis_download_dataset.R
\name{lm_nomis_download_dataset}
\alias{lm_nomis_download_dataset}
\title{Download a dataset from the NOMIS API}
\usage{
lm_nomis_download_dataset(
  base_uri,
  dataset_id,
  filter_string = NULL,
  row_limit = NULL
)
}
\arguments{
\item{base_uri}{\code{string}.
The base URI of the API.}

\item{dataset_id}{\code{string}.
An ID of an existing dataset in the NOMIS API. See \link{lm_nomis_datasets}.}

\item{filter_string}{\code{string}.
An optional string describing filter options and dimensions.}

\item{row_limit}{\code{string}.
An optional number of rows to return.}
}
\value{
\code{list}.
}
\description{
\code{lm_nomis_download_dataset} downloads a CSV file from the
   \href{https://www.nomisweb.co.uk/api/v01/help}{NOMIS API}.
}
\examples{
\dontrun{
lm_nomis_download_dataset(base_uri = "https://api.beta.ons.gov.uk/v1/", dataset_id = "NM_1_1")
lm_nomis_download_dataset(base_uri = "https://api.beta.ons.gov.uk/v1/", dataset_id = "NM_1_1", filter_string = "?geography=2092957697")
lm_nomis_download_dataset(base_uri = "https://api.beta.ons.gov.uk/v1/", dataset_id = "NM_1_1", row_limit = "5")
lm_nomis_download_dataset(base_uri = "https://api.beta.ons.gov.uk/v1/", dataset_id = "NM_1_1", filter_string = "?geography=2092957697", row_limit = "5")
}

}
