% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_ons_dataset_dimension_options.R
\name{lm_ons_dataset_dimension_options}
\alias{lm_ons_dataset_dimension_options}
\title{Return list of dimensions of a dataset in the ONS API}
\usage{
lm_ons_dataset_dimension_options(
  base_uri,
  dataset_id,
  dataset_edition,
  dataset_version,
  dataset_dimension
)
}
\arguments{
\item{base_uri}{\code{string}.
The base URI to connect to.
('datasets/' +  \code{dataset_id} + '/editions/' + \code{dataset_edition} + '/versions/' + \code{dataset_version} + '/dimensions/' + \code{dataset_dimension}, '/options'is appended in the function).}

\item{dataset_id}{\code{string}.
An ID of an existing dataset in the ONS API.}

\item{dataset_edition}{\code{string}.
An edition ID of an existing dataset in the ONS API.}

\item{dataset_version}{\code{string}.
A version ID of an existing dataset in the ONS API.}

\item{dataset_dimension}{\code{string}.
A name of an existing dimension in the specified dataset.}
}
\value{
\code{string} if one option available, \code{list} if multiple.
}
\description{
\code{lm_ons_dataset_dimensions} returns a list of dimensions of a dataset available from the ONS API.
}
\examples{
\dontrun{
lm_ons_dataset_dimension_options("https://api.beta.ons.gov.uk/v1/", "labour-market", "time-series", "2", "agegroups")
}

}
