% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_ons_datasets.R
\name{lm_ons_datasets}
\alias{lm_ons_datasets}
\title{Return list of datasets available from ONS API}
\usage{
lm_ons_datasets(base_uri, variables = NULL)
}
\arguments{
\item{base_uri}{\code{string}.
The base URI to connect to ('datasets' is appended in the function).}

\item{variables}{\code{list}.
Default = \code{null}.A list of desired variables to return.}
}
\value{
\code{list}.
}
\description{
\code{lm_ons_datasets} sends a GET request to a provided URI and formats the JSON response as an R list.
}
\examples{
\dontrun{
lm_ons_datasets("https://api.beta.ons.gov.uk/v1/", c("id", "description", "title"))
}

}
