% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_ons_download_dataset.R
\name{lm_ons_download_dataset}
\alias{lm_ons_download_dataset}
\title{Download a dataset from the ONS API}
\usage{
lm_ons_download_dataset(base_uri, filter_output_id)
}
\arguments{
\item{base_uri}{\code{string}.
The base URI of the API.}

\item{filter_output_id}{\code{string}.
A filter output ID. See \link{lm_ons_request_dataset}.}
}
\value{
\code{list}.
}
\description{
\code{lm_ons_download_dataset} returns the URI of a downloadable CSV file based on a
   \href{https://developer.ons.gov.uk/filter/}{filter} output ID created by
   \code{lm_ons_request_dataset}.
}
\examples{
\dontrun{
lm_ons_download_dataset(base_uri = "https://api.beta.ons.gov.uk/v1/", filter_output_id = "aa999999-8888-777b-c666-d5e555f44gh3")
}

}
