% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_ons_download_uri.R
\name{lm_ons_download_uri}
\alias{lm_ons_download_uri}
\title{Get the URI of a downloadable CSV file created in the ONS API}
\usage{
lm_ons_download_uri(base_uri, filter_output_id)
}
\arguments{
\item{base_uri}{\code{string}.
The base URI of the API.}

\item{filter_output_id}{\code{string}.
A filter output ID, created by \code{lm_ons_post_request}.}
}
\value{
\code{list}.
}
\description{
\code{lm_ons_download_uri} returns the URI of a downloadable CSV file based on a filter output ID created by \code{lm_ons_post_request}.
}
\examples{
\dontrun{
lm_ons_post_request(base_uri = "https://api.beta.ons.gov.uk/v1/", id = "labour-market", edition = "time-series", version = 18, dimension = NULL, option = NULL)
}

}
