% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_ons_post_request.R
\name{lm_ons_post_request}
\alias{lm_ons_post_request}
\title{Make POST request to ONS api and return dataset from CSV}
\usage{
lm_ons_post_request(
  base_uri,
  id,
  edition,
  version,
  dimension = NULL,
  option = NULL
)
}
\arguments{
\item{base_uri}{\code{string}.
The base URI used to determine the URIs to make GET and POST requests to.}

\item{id}{\code{string}.
The ID of a datasetin the ONS API.}

\item{edition}{\code{string}.
The edition of the dataset in the ONS API.}

\item{version}{\code{double}.
The version of the dataset in the ONS API.}

\item{dimension}{\code{string}.
(Default: \code{NULL}). The name of a dimension to filter on.}

\item{option}{\code{string}.
(Default: \code{NULL}). The option to use to filter by dimension.}
}
\value{
\code{list}.
}
\description{
\code{lm_format_get_request} creates a request body based on the \href{https://developer.ons.gov.uk/}{ONS API} structure
and sends it as a POST request to a provided URI.
The returned ID is then used to determine the URI where a CSV download has been created and a get request is sent,
with the resulting CSV being read into an R dataframe.
}
\examples{
\dontrun{
lm_ons_post_request(base_uri = "https://api.beta.ons.gov.uk/v1/", id = "labour-market", edition = "time-series", version = 18, dimension = NULL, option = NULL)
}

}
