% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_ons_dataset_editions.R
\name{lm_ons_dataset_editions}
\alias{lm_ons_dataset_editions}
\title{Return list of editions of a dataset in the ONS API}
\usage{
lm_ons_dataset_editions(base_uri, dataset_id)
}
\arguments{
\item{base_uri}{\code{string}.
The base URI to connect to ('datasets/' +  \code{dataset_id} + '/editions' is appended in the function).}

\item{dataset_id}{\code{string}.
An id of an existing dataset in the ONS API .}
}
\value{
\code{string} if one edition available, \code{list} if multiple.
}
\description{
\code{lm_ons_dataset_editions} returns a list of editions of a dataset available from the ONS API.
}
\examples{
\dontrun{
lm_ons_dataset_editions("https://api.beta.ons.gov.uk/v1/", "labour-market")
}

}
